requirements.txt
alembic==1.7.4
click==8.0.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.2
lark-parser==0.12.0
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
Pillow==8.4.0
SQLAlchemy==1.4.25
Werkzeug==2.0.2
zipp==3.6.0
